Refactor the DrawerToggle and Profile Card (#402)
* Refactor the drawer icon

* Lint

* Refactor ProfileCards
